+++ src/graphql.tsx
@@ -76,7 +76,10 @@ function getDisplayName(WrappedComponent) {
 // Helps track hot reloading.
 let nextVersion = 0;
 
-export function withApollo(WrappedComponent) {
+export function withApollo(
+  WrappedComponent,
+  operationOptions: OperationOption = {}
+) {
 
   const withDisplayName = `withApollo(${getDisplayName(WrappedComponent)})`;
 
@@ -100,10 +103,19 @@ export function withApollo(WrappedComponent) {
 
     }
 
+    getWrappedInstance() {
+      invariant(operationOptions.withRef,
+        `To access the wrapped instance, you need to specify ` +
+        `{ withRef: true } in the options`
+      );
+
+      return (this.refs as any).wrappedInstance;
+    }
 
     render() {
       const props = assign({}, this.props);
       props.client = this.client;
+      if (operationOptions.withRef) props.ref = 'wrappedInstance';
       return createElement(WrappedComponent, props);
     }
   }